clamp: move to a separate file
