+++ ggml/src/ggml-sycl/backend.hpp
@@ -36,6 +36,7 @@
 #include "getrows.hpp"
 #include "diagmask.hpp"
 #include "scale.hpp"
+#include "clamp.hpp"
 #include "gla.hpp"
 
 #endif // GGML_SYCL_BACKEND_HPP
